move display options from services to components
